app/scripts/requirements.txt
docopt
MySQL-python
sender
yagmail
requests
python-dateutil
